requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
decorator==5.0.9
distlib==0.3.2
filelock==3.0.12
idna==2.10
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
mysql-connector-python==8.0.25
neat-python==0.92
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pygame==2.0.1
Pygments==2.9.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==300
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.26.6
virtualenv==20.4.7
wcwidth==0.2.5
WebOb==1.8.7
xgboost==1.4.2
